fix: don't round down average gas per epoch
Also corrects the targetGps constant in the test harness. The constant
was chosen to be the expected Gps of the initial burst, causing the Gps
to stablize in the second part of the test at the floor.

Adding a check that the average measured gas price matches our desired
constant prevents this from silently passing.

+++ go/gas-oracle/gasprices/gas_price_updater.go
@@ -78,7 +78,7 @@ func (g *GasPriceUpdater) UpdateGasPrice() error {
 		totalGasUsed += gasUsed
 	}
 
-	averageGasPerSecond := float64(totalGasUsed / g.epochLengthSeconds)
+	averageGasPerSecond := float64(totalGasUsed) / float64(g.epochLengthSeconds)
 
 	log.Debug("UpdateGasPrice", "average-gas-per-second", averageGasPerSecond, "current-price", g.gasPricer.curPrice)
 	_, err = g.gasPricer.CompleteEpoch(averageGasPerSecond)
